Move buildVersionRange and update tests
