+++ vulnfeeds/cmd/cvelist2osv/version_extraction_test.go
@@ -32,65 +32,6 @@ func TestToVersionRangeType(t *testing.T) {
 	}
 }
 
-func TestBuildVersionRange(t *testing.T) {
-	tests := []struct {
-		name    string
-		intro   string
-		lastAff string
-		fixed   string
-		want    osvschema.Range
-	}{
-		{
-			name:  "intro and fixed",
-			intro: "1.0.0",
-			fixed: "1.0.1",
-			want: osvschema.Range{
-				Events: []osvschema.Event{
-					{Introduced: "1.0.0"},
-					{Fixed: "1.0.1"},
-				},
-			},
-		},
-		{
-			name:    "intro and last_affected",
-			intro:   "1.0.0",
-			lastAff: "1.0.0",
-			want: osvschema.Range{
-				Events: []osvschema.Event{
-					{Introduced: "1.0.0"},
-					{LastAffected: "1.0.0"},
-				},
-			},
-		},
-		{
-			name:  "only intro",
-			intro: "1.0.0",
-			want: osvschema.Range{
-				Events: []osvschema.Event{
-					{Introduced: "1.0.0"},
-				},
-			},
-		},
-		{
-			name: "empty intro",
-			want: osvschema.Range{
-				Events: []osvschema.Event{
-					{Introduced: "0"},
-				},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := buildVersionRange(tt.intro, tt.lastAff, tt.fixed)
-			if diff := cmp.Diff(tt.want, got); diff != "" {
-				t.Errorf("buildVersionRange() mismatch (-want +got):\n%s", diff)
-			}
-		})
-	}
-}
-
 func TestFindNormalAffectedRanges(t *testing.T) {
 	tests := []struct {
 		name          string
@@ -112,7 +53,7 @@ func TestFindNormalAffectedRanges(t *testing.T) {
 				},
 			},
 			wantRanges: []osvschema.Range{
-				buildVersionRange("1.0", "", "1.5"),
+				cves.BuildVersionRange("1.0", "", "1.5"),
 			},
 			wantRangeType: VersionRangeTypeSemver,
 		},
@@ -128,7 +69,7 @@ func TestFindNormalAffectedRanges(t *testing.T) {
 				},
 			},
 			wantRanges: []osvschema.Range{
-				buildVersionRange("0", "2.0", ""),
+				cves.BuildVersionRange("0", "2.0", ""),
 			},
 			wantRangeType: VersionRangeTypeSemver,
 		},
@@ -143,7 +84,7 @@ func TestFindNormalAffectedRanges(t *testing.T) {
 				},
 			},
 			wantRanges: []osvschema.Range{
-				buildVersionRange("2.0", "", "2.5"),
+				cves.BuildVersionRange("2.0", "", "2.5"),
 			},
 			wantRangeType: VersionRangeTypeEcosystem,
 		},
@@ -159,7 +100,7 @@ func TestFindNormalAffectedRanges(t *testing.T) {
 				},
 			},
 			wantRanges: []osvschema.Range{
-				buildVersionRange("deadbeef", "", ""),
+				cves.BuildVersionRange("deadbeef", "", ""),
 			},
 			wantRangeType: VersionRangeTypeGit,
 		},
@@ -228,7 +169,7 @@ func TestFindInverseAffectedRanges(t *testing.T) {
 			versionType: VersionRangeTypeSemver,
 			cnaAssigner: "Linux",
 			want: []osvschema.Range{
-				buildVersionRange("5.0.0", "", "5.10.1"),
+				cves.BuildVersionRange("5.0.0", "", "5.10.1"),
 			},
 		},
 		{
@@ -267,7 +208,7 @@ func TestFindInverseAffectedRanges(t *testing.T) {
 			versionType: VersionRangeTypeSemver,
 			cnaAssigner: "Linux",
 			want: []osvschema.Range{
-				buildVersionRange("4.0.0", "", "4.5.2"),
+				cves.BuildVersionRange("4.0.0", "", "4.5.2"),
 			},
 		},
 	}
