+++ vulnfeeds/cves/versions_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/osv/vulnfeeds/internal/testutils"
 	"github.com/google/osv/vulnfeeds/models"
+	"github.com/ossf/osv-schema/bindings/go/osvschema"
 )
 
 func loadTestData2(cveName string) Vulnerability {
@@ -1547,6 +1548,65 @@ func Test_MaybeUpdateVPRepoCache(t *testing.T) {
 			}
 			if !reflect.DeepEqual(tt.args.cache, tt.wantCache) {
 				t.Errorf("maybeUpdateVPRepoCache() have %#v, wanted %#v", tt.args.cache, tt.wantCache)
+			}
+		})
+	}
+}
+
+func TestBuildVersionRange(t *testing.T) {
+	tests := []struct {
+		name    string
+		intro   string
+		lastAff string
+		fixed   string
+		want    osvschema.Range
+	}{
+		{
+			name:  "intro and fixed",
+			intro: "1.0.0",
+			fixed: "1.0.1",
+			want: osvschema.Range{
+				Events: []osvschema.Event{
+					{Introduced: "1.0.0"},
+					{Fixed: "1.0.1"},
+				},
+			},
+		},
+		{
+			name:    "intro and last_affected",
+			intro:   "1.0.0",
+			lastAff: "1.0.0",
+			want: osvschema.Range{
+				Events: []osvschema.Event{
+					{Introduced: "1.0.0"},
+					{LastAffected: "1.0.0"},
+				},
+			},
+		},
+		{
+			name:  "only intro",
+			intro: "1.0.0",
+			want: osvschema.Range{
+				Events: []osvschema.Event{
+					{Introduced: "1.0.0"},
+				},
+			},
+		},
+		{
+			name: "empty intro",
+			want: osvschema.Range{
+				Events: []osvschema.Event{
+					{Introduced: "0"},
+				},
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := BuildVersionRange(tt.intro, tt.lastAff, tt.fixed)
+			if diff := cmp.Diff(tt.want, got); diff != "" {
+				t.Errorf("cves.BuildVersionRange() mismatch (-want +got):\n%s", diff)
 			}
 		})
 	}
